Extract CLI to its own module
As part of the effort to split up and test the code, I've extracted this
small module to handle CLI related things. It's far from pretty, but
much better covered, and decoupled.

Now, we can test how the parsing is working and also isolate the clap
related things into the awsudo::cli module.