Extract cache logic to its own module
Add a cache fetcher strategy for the Credentials. The Cache strategy
checks for the existence of the "~/.awsudo/$user", e.g.:
"~/.awsudo/staging" or "~/.awsudo/production". If the file exists and
meets the criterias for the caching, instead of requesting a new token,
the existing one is used instead, as a circuit-breaker, in the case of a
failure, e.g.: file not found, invalid format, missing attributes,
cache-expired..., the STS token is still triggered and generated.

At this point, there is no process for generating the cache file, which
should be done soon enough.